mc options to own file
